#121 changed log in functions again
